update kernel and related files for static quantization in FPGA v2 track

+++ src/fpga/V2/image.cpp
@@ -100,7 +100,7 @@ void concat_images(int8_t **images_in, float **scales_in, void *image_out,
         align_each_in_area_cw =
             align_to_x(channel_num[i] * width, IMAGE_ALIGNMENT);
         memcpy(
-            (int16_t *)image_out + tmp_channel +  // NOLINT
+            (int8_t *)image_out + tmp_channel +  // NOLINT
                 k * align_each_out_area_cw_differ,
             images_in_tmp[i] + j * channel_num[i] + k * align_each_in_area_cw,
             channel_num[i] * sizeof(int8_t));
